+++ common/src/main/java/com/fej1fun/potentials/capabilities/types/NoProviderBlockCapabilityHolder.java
@@ -15,8 +15,8 @@ public interface NoProviderBlockCapabilityHolder<X, Y> {
 
     @Nullable X getCapability(Level level, BlockPos pos, Y context);
     @Nullable X getCapability(Level level, BlockPos pos, BlockState state, BlockEntity blockEntity, Y context);
-    void registerForBlock(Supplier<Block> block);
-    void registerForBlockEntity(Supplier<BlockEntityType<?>> blockEntityType);
+    <T extends Block> void registerForBlock(Supplier<T> block);
+    <T extends BlockEntity> void registerForBlockEntity(Supplier<BlockEntityType<T>> blockEntityType);
     Identifier getIdentifier();
 
 }