+++ neoforge/src/main/java/com/fej1fun/potentials/neoforge/capabilities/holders/FluidBlockHolder.java
@@ -28,8 +28,8 @@ public class FluidBlockHolder implements NoProviderBlockCapabilityHolder<Univers
     public static final FluidBlockHolder INSTANCE = new FluidBlockHolder();
     private FluidBlockHolder() {registerSelf();}
 
-    private final Set<Supplier<Block>> registeredBlocks = new HashSet<>();
-    private final Set<Supplier<BlockEntityType<?>>> registeredBlockEntities = new HashSet<>();
+    private final Set<Supplier<? extends Block>> registeredBlocks = new HashSet<>();
+    private final Set<Supplier<? extends BlockEntityType<?>>> registeredBlockEntities = new HashSet<>();
 
     @Override
     public UniversalFluidStorage getCapability(Level level, BlockPos pos, Direction direction) {
@@ -43,12 +43,12 @@ public class FluidBlockHolder implements NoProviderBlockCapabilityHolder<Univers
     }
 
     @Override
-    public void registerForBlock(Supplier<Block> block) {
+    public <T extends Block> void registerForBlock(Supplier<T> block) {
         registeredBlocks.add(block);
     }
 
     @Override
-    public void registerForBlockEntity(Supplier<BlockEntityType<?>> blockEntityType) {
+    public <T extends BlockEntity> void registerForBlockEntity(Supplier<BlockEntityType<T>> blockEntityType) {
         registeredBlockEntities.add(blockEntityType);
     }